+++ src/api/axios/index.ts
@@ -1,1 +1,2 @@
+export * from './axios'
 export * from './axios-status'
