+++ src/api/endpoints/auth.ts
@@ -2,4 +2,5 @@ export const auth = {
   login: '/members/login',
   reissue: '/members/login/refresh',
   logout: '/members/logout',
+  check: '/members/check',
 }
