+++ src/api/fetches/auth.ts
@@ -1,6 +1,5 @@
 import { LoginRequest, LoginResponse, ReissueResponse } from '@/api/types'
-import { axiosStatus } from '@/api/utils'
-import { AXIOS, endpoints } from '@/api'
+import { AXIOS, endpoints, axiosStatus } from '@/api'
 
 export const auth = {
   login: async (request: LoginRequest): Promise<LoginResponse> => {
@@ -18,6 +17,13 @@ export const auth = {
   logout: async (): Promise<void> => {
     return axiosStatus(() => AXIOS.get<void>(endpoints.auth.logout), {
       onSuccess: (data) => data,
+    })
+  },
+
+  check: async (): Promise<boolean> => {
+    return axiosStatus(() => AXIOS.get<void>(endpoints.auth.check), {
+      onSuccess: () => true,
+      on401: () => false,
     })
   },
 }
